Better preselect symbol at cursor
We just pre-selected the nearest symbol.  Change that and select the
symbol at the top of the cursor.

If the cursor is on an empty line, first search for the next non-empty
line.  The idea here is to select the next symbol *iff* the user is
really just above the next function or class.  In other words: we tune
what "nearby"/"nearest symbol" means.