Retry fuzzy matching under certain circumstances
If we break because of a too high score, try again by matching the first
char at a later point.

This is a special case in this plugin, as we match against full line
contents.  I.e. a leading "c" could match at the BOL for "class ..." and
make the jump to the next char of the primer too wide.  But we don't
restrict this to the first char of `item`; let's generalize this for
the first match *if* it was a wide jump to a word boundary.  (Its score
becoming less than zero.)